Upgrade dcl (#7012)
* Upgrade dcl

* Add new basic test

* Simplify tests

* Fix yaml

* Use other zone

* Update tests

* Simplify tests

* Remove sample tests, use handwritten

* Rename test

* Test in separate project, add back docs-only sample

* Add test changes too

* Add a couple fields
