+++ .changelog/7012.txt
@@ -1,0 +1,3 @@
+```release-note:enhancement
+osconfig: added support for `skip_await_rollout` in `google_os_config_os_policy_assignment`
+```
